[AIML-ADC-9143] Fix issue where retrieving the application configuration fails for unauthenticated users. (#104)
While we updated to authorizer to allow users to fetch the application config without checking on their token we still had the authorizer associated with the resource. Unauthenticated users are missing a header required by the authorizer so we were returning a 401 before we even hit the lambda. This change adds support for registering 'unauthenticated' endpoints and uses that for the GET app-config resource.