[AIML-ADC-9143] Include authorization header when fetching configuration (#109)
* [AIML-ADC-9143] Fix issue where retrieving the application configuration fails for unauthenticated users.
While we updated to authorizer to allow users to fetch the application config without checking on their token we still had the authorizer associated with the resource. Unauthenticated users are missing a header required by the authorizer so we were returning a 401 before we even hit the lambda. This change updates the axios utils to always include the authorization header even if the user does not have a valid token. This ensures the authorizer is always hit.

* Update notebook reducer tests to ensure proper token is being included with requests.

---------

Co-authored-by: LightSeekerSC <46975522+LightSeekerSC@users.noreply.github.com>